+++ CHANGES.rst
@@ -27,7 +27,9 @@
 - Turned off GMOS-S binning restriction
 - Add GTC OSIRIS spectrograph
 - Updates for docs on adding new spectrographs.  And a bok test
-
+- Added a new ``pypeit_collate_1d`` tool to automatically group 1D Spectra from
+  multiple files by group and coadd them.
+- PypeIt will now add HISTORY keyword entries to FITS files.
 
 1.3.3 (24 Feb 2021)
 -------------------